requirements.txt
pillow
numpy
pandas
opencv-python
opencv-contrib-python
matplotlib
tensorflow
google-generativeai
python-dotenv
pillow
